+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add "cdtb" in RackAttack logs to ease finding them.
 - Optimize Users::Remover task, retrieve reporter_user only once
 - Update way of retrieving manifest form attributes on census task
+- Check zeitwerk class loading.
 
 ## [0.5.2] - 2025-05-23 (patch - Sense por ni flotador)